% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca_cep.R
\name{busca_cep}
\alias{busca_cep}
\title{Busca por CEP}
\usage{
busca_cep(cep = "01001000", token = NULL)
}
\arguments{
\item{cep}{CEP (postal code), \code{character}.}

\item{token}{Token de autorização. Veja <http://cepaberto.com/users/register>.}
}
\description{
Busca por CEP (search for information by postal code).
}
