% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca_cidade.R, R/busca_estado_cidade.R
\name{busca_cidade}
\alias{busca_cidade}
\title{Busca Bairros por Estado}
\usage{
busca_cidade(
  estado = c("AC", "AL", "AP", "AM", "BA", "CE", "DF", "ES", "GO", "MA", "MT", "MS",
    "MG", "PA", "PB", "PR", "PE", "PI", "RJ", "RN", "RS", "RO", "RR", "SC", "SP", "SE",
    "TO"),
  cidade = "",
  token = NULL
)

busca_cidade(
  estado = c("AC", "AL", "AP", "AM", "BA", "CE", "DF", "ES", "GO", "MA", "MT", "MS",
    "MG", "PA", "PB", "PR", "PE", "PI", "RJ", "RN", "RS", "RO", "RR", "SC", "SP", "SE",
    "TO"),
  cidade = "",
  token = NULL
)
}
\arguments{
\item{estado}{sigla do estado (acronym of the state).}

\item{cidade}{\code{character}. Nome da cidade, por exemplo, \code{"Salvador"}. (Name of the city.)}

\item{token}{Token de autorização. Veja <http://cepaberto.com/users/register>.}
}
\description{
Busca bairros por estado.

Busca bairros por estado.
}
\examples{
\dontrun{
ubatuba <- busca_cidade(estado = "SP", cidade = "Ubatuba", token = token)
}
\dontrun{
ubatuba <- busca_cidade(estado = "SP", cidade = "Ubatuba", token = token)
}
}
