% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca_cidades.R
\name{busca_cidades}
\alias{busca_cidades}
\title{Busca Cidades por Estado}
\usage{
busca_cidades(
  estado = c("AC", "AL", "AP", "AM", "BA", "CE", "DF", "ES", "GO", "MA", "MT", "MS",
    "MG", "PA", "PB", "PR", "PE", "PI", "RJ", "RN", "RS", "RO", "RR", "SC", "SP", "SE",
    "TO"),
  token = NULL
)
}
\arguments{
\item{estado}{sigla do estado (acronym of the state).}

\item{token}{Token de autorização. Veja <http://cepaberto.com/users/register>.}
}
\description{
Busca cidades por estado.
}
\examples{
\dontrun{
ubatuba <- busca_cidades(estado = "AM", token = token)
}
}
