% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca_estado.R
\name{busca_estado}
\alias{busca_estado}
\title{Busca Bairros por Estado}
\usage{
busca_estado(estado = c("AC", "AL", "AP", "AM", "BA", "CE", "DF", "ES", "GO",
  "MA", "MT", "MS", "MG", "PA", "PB", "PR", "PE", "PI", "RJ", "RN", "RS", "RO",
  "RR", "SC", "SP", "SE", "TO"), token = NULL)
}
\arguments{
\item{estado}{sigla do estado.}

\item{token}{Token de autorização. Veja <http://cepaberto.com/users/register>.}
}
\description{
Busca bairros por estado.
}
