% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca_latlon.R
\name{busca_latlon}
\alias{busca_latlon}
\title{Busca Bairros por Latitude e Longitude}
\usage{
busca_latlon(lat = NULL, long = NULL, token = NULL)
}
\arguments{
\item{lat}{\code{numerico}. Latitude}

\item{long}{\code{numerico}. Longitude.}

\item{token}{Token de autorização. Veja <http://cepaberto.com/users/register>}
}
\description{
Busca bairros por latitude e longitude.
}
\examples{
\dontrun{
busca_latlon(lat = -20.55, long = -43.63, token = XXXXXX)
}
}
