% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca_multi.R
\name{busca_multi}
\alias{busca_multi}
\title{Busca por CEP}
\usage{
busca_multi(lista_ceps = c("01001000", "01001001"), token = NULL)
}
\arguments{
\item{lista_ceps}{lista de CEPs (postal code list).}

\item{token}{Token de autorização. Veja <http://cepaberto.com/users/register>.}
}
\description{
Busca usando multiplos CEP (search for information using multiple postal code).
}
