% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\docType{data}
\name{sp}
\alias{sp}
\title{Base de CEP, endereços e bairros de São Paulo}
\format{tibble de 258765 linhas, 7 colunas.}
\usage{
sp
}
\value{
\itemize{
 \item{\code{cep}}{}
 \item{\code{local}}{}
 \item{\code{bairro}}{}
 \item{\code{id_cidade: }}{id da cidade na base do API de CEPaberto.}
 \item{\code{id_estado: }}{id do estado na base do API de CEPaberto.}
 \item{\code{cidade}}{}
 \item{\code{estado}}{}
}
}
\description{
This is a dataset of the nominal votes in the Brazilian
Federal Senate, from all those available on the API from 1991 onwards.
}
\keyword{datasets}
